new change with db core

+++ xianchangjia/DAOModel/coredata/LXUser.h
@@ -23,7 +23,7 @@
 @property (nonatomic, strong, readonly) NSString *phone;
 @property (nonatomic, assign, readonly) int sex;
 @property (nonatomic, assign, readonly) int height;
-@property (nonatomic, assign, readonly) int create_time;
+@property (nonatomic, assign, readonly) double create_time;
 @property (nonatomic, assign, readonly) int active_by;
 @property (nonatomic, assign, readonly) int active_level;
 @property (nonatomic, assign, readonly) int actor;
